# BVH optimization for SignedDistance