# Summary

This PR adds an optimization for point queries against a BVH with the device traversal API. At a given node in the BVH, when both child nodes need to be traversed, we first traverse the node with the bounding box that is closer to the query point. This improves the performance of signed-distance queries, since the current-minimum distance for a given query point converges much faster to the actual minimum distance. As a result, more of the tree can be excluded from traversal.

This is not added to the `BVH::findCandidates` methods, because in that case, we must traverse all tentatively-intersecting nodes regardless.

## Performance

Measured with `./examples/quest_signed_distance_interface_ex --batched --box_dims 200 200 200 --file ../data/quest/plane.stl`

|          | `develop` | this branch | speedup |
| ------ | ---------- | ------------ | --------- |
| rzgenie | 109.177s | 29.8469s | 3.66x |
| rzansel | 9.75109s | 3.3926s | 2.87x |